Breadcrumbs
The idea here is that we remove the poller from the querier.  To do so,
we need to pass the blocklist from the frontend to the querier through
the job interface.  On the surface this seems reasonable.

Note that currently in the querier we consult the entire blocklist and
determine which blocks belong to the query.  The frontend could do this
processing, so that only the block which are sent through the job
interface are ones that matched.  This will increase the burden on the
frontend, but unsure how much.

Additional considerations here, are that we will still likely need to
round-trip the proto blockmeta to go blockmeta on each query to send
through the job interface, which doesn't sound like an improvement over
the current maintenance of the blocklist, but would probably need to be
completed and tested before we know for sure.  It would also be subject
to the query patterns of the tenant, etc.